Drasil.GOOL.LanguageRenderer.CommonPseudoOO

Description

Implementations defined here are valid in some, but not all, language renderers

Documentation

int :: RenderSym r => VSType r

constructor :: RenderSym r => Label -> [MSParameter r] -> Initializers r -> MSBody r -> SMethod r

doxFunc :: RenderSym r => String -> [String] -> Maybe String -> SMethod r -> SMethod r

doxClass :: RenderSym r => String -> SClass r -> SClass r

doxMod :: RenderSym r => String -> String -> [String] -> String -> SFile r -> SFile r

docMod' :: RenderSym r => String -> String -> [String] -> String -> SFile r -> SFile r

extVar :: RenderSym r => Label -> Label -> VSType r -> SVariable r

classVar :: RenderSym r => (Doc -> Doc -> Doc) -> VSType r -> SVariable r -> SVariable r

objVarSelf :: RenderSym r => SVariable r -> SVariable r

indexOf :: RenderSym r => Label -> SValue r -> SValue r -> SValue r

listAddFunc :: RenderSym r => Label -> SValue r -> SValue r -> VSFunction r

discardFileLine :: RenderSym r => Label -> SValue r -> MSStatement r

intClass :: (RenderSym r, Monad r) => (Label -> Doc -> Doc -> Doc -> Doc -> Doc) -> Label -> r (Scope r) -> r ParentSpec -> [CSStateVar r] -> [SMethod r] -> [SMethod r] -> CS (r Doc)

An internal function for creating a class. Parameters: render function, class name, scope, parent, class variables, constructor(s), methods

funcType :: RenderSym r => [VSType r] -> VSType r -> VSType r

buildModule :: RenderSym r => Label -> FS Doc -> FS Doc -> FS Doc -> [SMethod r] -> [SClass r] -> FSModule r

arrayType :: RenderSym r => VSType r -> VSType r

pi :: RenderSym r => SValue r

printSt :: RenderSym r => SValue r -> SValue r -> MSStatement r

arrayDec :: RenderSym r => SValue r -> SVariable r -> MSStatement r

arrayDecDef :: RenderSym r => SVariable r -> [SValue r] -> MSStatement r

openFileA :: RenderSym r => (SValue r -> VSType r -> SValue r -> SValue r) -> SVariable r -> SValue r -> MSStatement r

forEach :: RenderSym r => Doc -> Doc -> Doc -> Doc -> SVariable r -> SValue r -> MSBody r -> MSStatement r

docMain :: RenderSym r => MSBody r -> SMethod r

mainFunction :: RenderSym r => VSType r -> Label -> MSBody r -> SMethod r

buildModule' :: RenderSym r => Label -> (String -> r (Import r)) -> [Label] -> [SMethod r] -> [SClass r] -> FSModule r

Used by the language renderers to build the module. n is the module name inc is the include is is the import statements ms is the class methods cs is the classes

call' :: RenderSym r => String -> Maybe Library -> Maybe Doc -> MixedCall r

First parameter is language name, rest similar to call from RendererClasses

listSizeFunc :: RenderSym r => VSFunction r

listAccessFunc' :: RenderSym r => Label -> VSType r -> SValue r -> VSFunction r

string :: RenderSym r => VSType r

constDecDef :: RenderSym r => SVariable r -> SValue r -> MSStatement r

docInOutFunc :: RenderSym r => ([SVariable r] -> [SVariable r] -> [SVariable r] -> MSBody r -> SMethod r) -> String -> [(String, SVariable r)] -> [(String, SVariable r)] -> [(String, SVariable r)] -> MSBody r -> SMethod r

bindingError :: String -> String

extFuncAppMixedArgs :: RenderSym r => Library -> MixedCall r

notNull :: RenderSym r => String -> SValue r -> SValue r

listDecDef :: RenderSym r => SVariable r -> [SValue r] -> MSStatement r

destructorError :: String -> String

stateVarDef :: (RenderSym r, Monad r) => r (Scope r) -> r (Permanence r) -> SVariable r -> SValue r -> CS (r Doc)

constVar :: (RenderSym r, Monad r) => Doc -> r (Scope r) -> SVariable r -> SValue r -> CS (r Doc)

litArray :: RenderSym r => (Doc -> Doc) -> VSType r -> [SValue r] -> SValue r

listSetFunc :: RenderSym r => (Doc -> Doc -> Doc) -> SValue r -> SValue r -> SValue r -> VSFunction r

extraClass :: RenderSym r => Label -> Maybe Label -> [CSStateVar r] -> [SMethod r] -> [SMethod r] -> SClass r

listAccessFunc :: RenderSym r => VSType r -> SValue r -> VSFunction r

doubleRender :: String

double :: RenderSym r => VSType r

openFileR :: RenderSym r => (SValue r -> VSType r -> SValue r) -> SVariable r -> SValue r -> MSStatement r

openFileW :: RenderSym r => (SValue r -> VSType r -> SValue r -> SValue r) -> SVariable r -> SValue r -> MSStatement r

stateVar :: (RenderSym r, Monad r) => r (Scope r) -> r (Permanence r) -> SVariable r -> CS (r Doc)

self :: RenderSym r => SVariable r

multiAssign :: RenderSym r => (Doc -> Doc) -> [SVariable r] -> [SValue r] -> MSStatement r

multiReturn :: RenderSym r => (Doc -> Doc) -> [SValue r] -> MSStatement r

listDec :: RenderSym r => SVariable r -> MSStatement r

funcDecDef :: RenderSym r => SVariable r -> [SVariable r] -> MSBody r -> MSStatement r

inOutCall :: RenderSym r => (Label -> VSType r -> [SValue r] -> SValue r) -> Label -> [SValue r] -> [SVariable r] -> [SVariable r] -> MSStatement r

forLoopError :: String -> String

mainBody :: RenderSym r => MSBody r -> SMethod r

inOutFunc :: RenderSym r => (VSType r -> [MSParameter r] -> MSBody r -> SMethod r) -> [SVariable r] -> [SVariable r] -> [SVariable r] -> MSBody r -> SMethod r

docInOutFunc' :: RenderSym r => FuncDocRenderer -> ([SVariable r] -> [SVariable r] -> [SVariable r] -> MSBody r -> SMethod r) -> String -> [(String, SVariable r)] -> [(String, SVariable r)] -> [(String, SVariable r)] -> MSBody r -> SMethod r

boolRender :: String

bool :: RenderSym r => VSType r

floatRender :: String

float :: RenderSym r => VSType r

stringRender' :: String

string' :: RenderSym r => VSType r

inherit :: Monad r => Maybe Label -> r ParentSpec

implements :: Monad r => [Label] -> r ParentSpec

listSize :: RenderSym r => SValue r -> SValue r

Call to get the size of a list in a language where this is not a method.

listAdd :: RenderSym r => SValue r -> SValue r -> SValue r -> SValue r

Call to insert a value into a list in a language where this is not a method.

listAppend :: RenderSym r => SValue r -> SValue r -> SValue r

Call to append a value to a list in a language where this is not a method.

intToIndex :: SValue r -> SValue r

Convert an integer to an index in a 0-indexed language Since GOOL is 0-indexed, no adjustments need be made

indexToInt :: SValue r -> SValue r

Convert an index to an integer in a 0-indexed language Since GOOL is 0-indexed, no adjustments need be made

intToIndex' :: RenderSym r => SValue r -> SValue r

Convert an integer to an index in a 1-indexed language Since GOOL is 0-indexed, we need to add 1

indexToInt' :: RenderSym r => SValue r -> SValue r

Convert an index to an integer in a 1-indexed language Since GOOL is 0-indexed, we need to subtract 1