Drasil.GOOL.RendererClassesCommon

Documentation

class (AssignStatement r, DeclStatement r, IOStatement r, StringStatement r, FuncAppStatement r, CommentStatement r, ControlStatement r, Argument r, Literal r, MathConstant r, VariableValue r, CommandLineArgs r, NumericExpression r, BooleanExpression r, Comparison r, List r, InternalList r, VectorExpression r, TypeElim r, VariableElim r, RenderBlock r, BlockElim r, RenderBody r, BodyElim r, InternalListFunc r, RenderFunction r, FunctionElim r, OpElim r, RenderParam r, ParamElim r, RenderVisibility r, VisibilityElim r, InternalAssignStmt r, InternalIOStmt r, InternalControlStmt r, RenderStatement r, StatementElim r, RenderType r, InternalTypeElim r, RenderValue r, ValueElim r, RenderVariable r, InternalVarElim r, ImportSym r, ImportElim r, UnaryOpSym r, BinaryOpSym r, BlockCommentSym r, BlockCommentElim r, ValueExpression r, RenderMethod r, MethodElim r, ParameterSym r, ScopeElim r) => CommonRenderSym r

class ImportSym r where

Associated Types

type Import r

Methods

langImport :: Label -> r (Import r)

modImport :: Label -> r (Import r)

class ImportElim r where

Methods

import' :: r (Import r) -> Doc

class RenderBody r where

Methods

multiBody :: [MSBody r] -> MSBody r

class BodyElim r where

Methods

body :: r (Body r) -> Doc

class RenderBlock r where

Methods

multiBlock :: [MSBlock r] -> MSBlock r

class BlockElim r where

Methods

block :: r (Block r) -> Doc

class RenderType r where

Methods

multiType :: [VSType r] -> VSType r

typeFromData :: CodeType -> String -> Doc -> VSType r

class InternalTypeElim r where

Methods

type' :: r (Type r) -> Doc

type VSUnOp a = VS (a (UnaryOp a))

class UnaryOpSym r where

Associated Types

type UnaryOp r

Methods

notOp :: VSUnOp r

negateOp :: VSUnOp r

sqrtOp :: VSUnOp r

absOp :: VSUnOp r

logOp :: VSUnOp r

lnOp :: VSUnOp r

expOp :: VSUnOp r

sinOp :: VSUnOp r

cosOp :: VSUnOp r

tanOp :: VSUnOp r

asinOp :: VSUnOp r

acosOp :: VSUnOp r

atanOp :: VSUnOp r

floorOp :: VSUnOp r

ceilOp :: VSUnOp r

type VSBinOp a = VS (a (BinaryOp a))

class BinaryOpSym r where

Associated Types

type BinaryOp r

Methods

equalOp :: VSBinOp r

notEqualOp :: VSBinOp r

greaterOp :: VSBinOp r

greaterEqualOp :: VSBinOp r

lessOp :: VSBinOp r

lessEqualOp :: VSBinOp r

plusOp :: VSBinOp r

minusOp :: VSBinOp r

multOp :: VSBinOp r

divideOp :: VSBinOp r

powerOp :: VSBinOp r

moduloOp :: VSBinOp r

andOp :: VSBinOp r

orOp :: VSBinOp r

class OpElim r where

Methods

uOp :: r (UnaryOp r) -> Doc

bOp :: r (BinaryOp r) -> Doc

uOpPrec :: r (UnaryOp r) -> Int

bOpPrec :: r (BinaryOp r) -> Int

class RenderVariable r where

Methods

varFromData :: Binding -> String -> VSType r -> Doc -> SVariable r

class InternalVarElim r where

Methods

variableBind :: r (Variable r) -> Binding

variable :: r (Variable r) -> Doc

class RenderValue r where

Methods

inputFunc :: SValue r

printFunc :: SValue r

printLnFunc :: SValue r

printFileFunc :: SValue r -> SValue r

printFileLnFunc :: SValue r -> SValue r

cast :: VSType r -> SValue r -> SValue r

call :: Maybe Library -> Maybe Doc -> MixedCall r

Very generic internal function for generating calls, to reduce repeated code throughout generators. Parameters are: maybe name of external module, maybe Doc for object variable (including separator between object and function) for method calls.

valFromData :: Maybe Int -> Maybe Integer -> VSType r -> Doc -> SValue r

class ValueElim r where

Methods

valuePrec :: r (Value r) -> Maybe Int

valueInt :: r (Value r) -> Maybe Integer

value :: r (Value r) -> Doc

class InternalListFunc r where

Methods

listSizeFunc :: SValue r -> VSFunction r

List

listAddFunc :: SValue r -> SValue r -> SValue r -> VSFunction r

List, Index, Value

listAppendFunc :: SValue r -> SValue r -> VSFunction r

List, Value

listAccessFunc :: VSType r -> SValue r -> VSFunction r

List, Index

listSetFunc :: SValue r -> SValue r -> SValue r -> VSFunction r

List, Index, Value

class RenderFunction r where

Methods

funcFromData :: Doc -> VSType r -> VSFunction r

class FunctionElim r where

Methods

functionType :: r (Function r) -> r (Type r)

function :: r (Function r) -> Doc

class InternalAssignStmt r where

Methods

multiAssign :: [SVariable r] -> [SValue r] -> MSStatement r

class InternalIOStmt r where

Methods

printSt :: Bool -> Maybe (SValue r) -> SValue r -> SValue r -> MSStatement r

class InternalControlStmt r where

Methods

multiReturn :: [SValue r] -> MSStatement r

class RenderStatement r where

Methods

stmt :: MSStatement r -> MSStatement r

loopStmt :: MSStatement r -> MSStatement r

stmtFromData :: Doc -> Terminator -> MSStatement r

class StatementElim r where

Methods

statement :: r (Statement r) -> Doc

statementTerm :: r (Statement r) -> Terminator

class RenderVisibility r where

Methods

visibilityFromData :: VisibilityTag -> Doc -> r (Visibility r)

class VisibilityElim r where

Methods

visibility :: r (Visibility r) -> Doc

type MSMthdType a = MS (a (MethodType a))

class TypeSym r => MethodTypeSym r where

Associated Types

type MethodType r

Methods

mType :: VSType r -> MSMthdType r

class RenderParam r where

Methods

paramFromData :: SVariable r -> Doc -> MSParameter r

class ParamElim r where

Methods

parameterName :: r (Parameter r) -> Label

parameterType :: r (Parameter r) -> r (Type r)

parameter :: r (Parameter r) -> Doc

class (MethodTypeSym r, BlockCommentSym r) => RenderMethod r where

Methods

commentedFunc :: MS (r (BlockComment r)) -> SMethod r -> SMethod r

Takes a BlockComment and a method and generates a function.

mthdFromData :: VisibilityTag -> Doc -> SMethod r

class MethodElim r where

Methods

method :: r (Method r) -> Doc

class BlockCommentSym r where

Associated Types

type BlockComment r

Methods

blockComment :: [String] -> r (BlockComment r)

docComment :: State a [String] -> State a (r (BlockComment r))

Converts a list of strings into a block comment

class BlockCommentElim r where

Methods

blockComment' :: r (BlockComment r) -> Doc

class ScopeElim r where

Methods

scopeData :: r (Scope r) -> ScopeData