Drasil.Shared.InterfaceCommon

Documentation

type Label = String

type Library = String

type MSBody a = MS (a (Body a))

type MSBlock a = MS (a (Block a))

type VSFunction a = VS (a (Function a))

type VSType a = VS (a (Type a))

type SVariable a = VS (a (Variable a))

type SValue a = VS (a (Value a))

type VSThunk a = VS (a (Thunk a))

type MSStatement a = MS (a (Statement a))

type MSParameter a = MS (a (Parameter a))

type SMethod a = MS (a (Method a))

type NamedArgs r = [(SVariable r, SValue r)]

type MixedCall r = Label -> VSType r -> [SValue r] -> NamedArgs r -> SValue r

type MixedCtorCall r = VSType r -> [SValue r] -> NamedArgs r -> SValue r

type PosCall r = Label -> VSType r -> [SValue r] -> SValue r

type PosCtorCall r = VSType r -> [SValue r] -> SValue r

type InOutCall r = Label -> [SValue r] -> [SVariable r] -> [SVariable r] -> MSStatement r

type InOutFunc r = [SVariable r] -> [SVariable r] -> [SVariable r] -> MSBody r -> SMethod r

type DocInOutFunc r = String -> [(String, SVariable r)] -> [(String, SVariable r)] -> [(String, SVariable r)] -> MSBody r -> SMethod r

class (VectorType r, VectorDecl r, VectorThunk r, VectorExpression r, ThunkAssign r, AssignStatement r, DeclStatement r, IOStatement r, StringStatement r, FunctionSym r, FuncAppStatement r, CommentStatement r, ControlStatement r, InternalList r, Argument r, Literal r, MathConstant r, VariableValue r, CommandLineArgs r, NumericExpression r, BooleanExpression r, Comparison r, ValueExpression r, List r, Set r, TypeElim r, VariableElim r, MethodSym r, ScopeSym r) => SharedProg r

class BlockSym r => BodySym r where

Associated Types

type Body r

Methods

body :: [MSBlock r] -> MSBody r

addComments :: Label -> MSBody r -> MSBody r

bodyStatements :: BodySym r => [MSStatement r] -> MSBody r

oneLiner :: BodySym r => MSStatement r -> MSBody r

class StatementSym r => BlockSym r where

Associated Types

type Block r

Methods

block :: [MSStatement r] -> MSBlock r

class TypeSym r where

Associated Types

type Type r

Methods

bool :: VSType r

int :: VSType r

float :: VSType r

double :: VSType r

char :: VSType r

string :: VSType r

infile :: VSType r

outfile :: VSType r

listType :: VSType r -> VSType r

setType :: VSType r -> VSType r

arrayType :: VSType r -> VSType r

listInnerType :: VSType r -> VSType r

funcType :: [VSType r] -> VSType r -> VSType r

void :: VSType r

class TypeSym r => TypeElim r where

Methods

getType :: r (Type r) -> CodeType

getTypeString :: r (Type r) -> String

class TypeSym r => VariableSym r where

Associated Types

type Variable r

Methods

var :: Label -> VSType r -> SVariable r

constant :: Label -> VSType r -> SVariable r

extVar :: Library -> Label -> VSType r -> SVariable r

arrayElem :: Integer -> SVariable r -> SVariable r

class ScopeSym r where

Associated Types

type Scope r

Methods

global :: r (Scope r)

mainFn :: r (Scope r)

local :: r (Scope r)

convScope :: ScopeSym r => ScopeData -> r (Scope r)

class VariableSym r => VariableElim r where

Methods

variableName :: r (Variable r) -> String

variableType :: r (Variable r) -> r (Type r)

listOf :: VariableSym r => Label -> VSType r -> SVariable r

listVar :: VariableSym r => Label -> VSType r -> SVariable r

class TypeSym r => ValueSym r where

Associated Types

type Value r

Methods

valueType :: r (Value r) -> r (Type r)

class ValueSym r => Argument r where

Methods

pointerArg :: SValue r -> SValue r

class ValueSym r => Literal r where

Methods

litTrue :: SValue r

litFalse :: SValue r

litChar :: Char -> SValue r

litDouble :: Double -> SValue r

litFloat :: Float -> SValue r

litInt :: Integer -> SValue r

litString :: String -> SValue r

litArray :: VSType r -> [SValue r] -> SValue r

litList :: VSType r -> [SValue r] -> SValue r

litSet :: VSType r -> [SValue r] -> SValue r

litZero :: (TypeElim r, Literal r) => VSType r -> SValue r

class ValueSym r => MathConstant r where

Methods

pi :: SValue r

class (VariableSym r, ValueSym r) => VariableValue r where

Methods

valueOf :: SVariable r -> SValue r

class ValueSym r => CommandLineArgs r where

Methods

arg :: Integer -> SValue r

argsList :: SValue r

argExists :: Integer -> SValue r

class ValueSym r => NumericExpression r where

Methods

(#~) :: SValue r -> SValue r infixl 8

(#/^) :: SValue r -> SValue r infixl 7

(#|) :: SValue r -> SValue r infixl 7

(#+) :: SValue r -> SValue r -> SValue r infixl 5

(#-) :: SValue r -> SValue r -> SValue r infixl 5

(#*) :: SValue r -> SValue r -> SValue r infixl 6

(#/) :: SValue r -> SValue r -> SValue r infixl 6

(#%) :: SValue r -> SValue r -> SValue r infixl 6

(#^) :: SValue r -> SValue r -> SValue r infixl 7

log :: SValue r -> SValue r

ln :: SValue r -> SValue r

exp :: SValue r -> SValue r

sin :: SValue r -> SValue r

cos :: SValue r -> SValue r

tan :: SValue r -> SValue r

csc :: SValue r -> SValue r

sec :: SValue r -> SValue r

cot :: SValue r -> SValue r

arcsin :: SValue r -> SValue r

arccos :: SValue r -> SValue r

arctan :: SValue r -> SValue r

floor :: SValue r -> SValue r

ceil :: SValue r -> SValue r

class ValueSym r => BooleanExpression r where

Methods

(?!) :: SValue r -> SValue r infixr 6

(?&&) :: SValue r -> SValue r -> SValue r infixl 2

(?||) :: SValue r -> SValue r -> SValue r infixl 1

class ValueSym r => Comparison r where

Methods

(?<) :: SValue r -> SValue r -> SValue r infixl 4

(?<=) :: SValue r -> SValue r -> SValue r infixl 4

(?>) :: SValue r -> SValue r -> SValue r infixl 4

(?>=) :: SValue r -> SValue r -> SValue r infixl 4

(?==) :: SValue r -> SValue r -> SValue r infixl 3

(?!=) :: SValue r -> SValue r -> SValue r infixl 3

class (VariableSym r, ValueSym r) => ValueExpression r where

Methods

inlineIf :: SValue r -> SValue r -> SValue r -> SValue r

funcAppMixedArgs :: MixedCall r

extFuncAppMixedArgs :: Library -> MixedCall r

libFuncAppMixedArgs :: Library -> MixedCall r

lambda :: [SVariable r] -> SValue r -> SValue r

notNull :: SValue r -> SValue r

funcApp :: ValueExpression r => PosCall r

funcAppNamedArgs :: ValueExpression r => Label -> VSType r -> NamedArgs r -> SValue r

extFuncApp :: ValueExpression r => Library -> PosCall r

libFuncApp :: ValueExpression r => Library -> PosCall r

exists :: ValueExpression r => SValue r -> SValue r

class ValueSym r => List r where

Methods

intToIndex :: SValue r -> SValue r

Does any necessary conversions from GOOL's zero-indexed assumptions to the target language's assumptions

indexToInt :: SValue r -> SValue r

Does any necessary conversions from the target language's indexing assumptions assumptions to GOOL's zero-indexed assumptions

listSize :: SValue r -> SValue r

Finds the size of a list. Arguments are: List

listAdd :: SValue r -> SValue r -> SValue r -> SValue r

Inserts a value into a list. Arguments are: List, Index, Value

listAppend :: SValue r -> SValue r -> SValue r

Appens a value to a list. Arguments are: List, Value

listAccess :: SValue r -> SValue r -> SValue r

Gets the value of an index of a list. Arguments are: List, Index

listSet :: SValue r -> SValue r -> SValue r -> SValue r

Sets the value of an index of a list. Arguments are: List, Index, Value

indexOf :: SValue r -> SValue r -> SValue r

Finds the index of the first occurrence of a value in a list. Arguments are: List, Value

class ValueSym r => Set r where

Methods

contains :: SValue r -> SValue r -> SValue r

Checks membership Arguments are: Set, Value

setAdd :: SValue r -> SValue r -> SValue r

Inserts a value into a set Arguments are: Set, Value

setRemove :: SValue r -> SValue r -> SValue r

Removes a value from a set Arguments are: Set, Value

setUnion :: SValue r -> SValue r -> SValue r

Removes a value from a set Arguments are: Set, Set

class ValueSym r => InternalList r where

Methods

listSlice' :: Maybe (SValue r) -> Maybe (SValue r) -> Maybe (SValue r) -> SVariable r -> SValue r -> MSBlock r

listSlice :: InternalList r => SVariable r -> SValue r -> Maybe (SValue r) -> Maybe (SValue r) -> Maybe (SValue r) -> MSBlock r

Creates a slice of a list and assigns it to a variable. Arguments are: Variable to assign List to read from (optional) Start index inclusive. (if Nothing, then list start if step > 0, list end if step < 0) (optional) End index exclusive. (if Nothing, then list end if step > 0, list start if step > 0) (optional) Step (if Nothing, then defaults to 1)

listIndexExists :: (List r, Comparison r) => SValue r -> SValue r -> SValue r

at :: List r => SValue r -> SValue r -> SValue r

class ThunkSym r

Associated Types

type Thunk (r :: Type -> Type)

class TypeSym r => VectorType r where

Methods

vecType :: VSType r -> VSType r

class DeclStatement r => VectorDecl r where

Methods

vecDec :: Integer -> SVariable r -> r (Scope r) -> MSStatement r

vecDecDef :: SVariable r -> r (Scope r) -> [SValue r] -> MSStatement r

class (VariableSym r, ThunkSym r) => VectorThunk r where

Methods

vecThunk :: SVariable r -> VSThunk r

class (ThunkSym r, ValueSym r) => VectorExpression r where

Methods

vecScale :: SValue r -> VSThunk r -> VSThunk r

vecAdd :: VSThunk r -> VSThunk r -> VSThunk r

vecIndex :: SValue r -> VSThunk r -> SValue r

vecDot :: VSThunk r -> VSThunk r -> VSThunk r

class (VariableSym r, ThunkSym r, StatementSym r) => ThunkAssign r where

Methods

thunkAssign :: SVariable r -> VSThunk r -> MSStatement r

class ValueSym r => StatementSym r where

Associated Types

type Statement r

Methods

valStmt :: SValue r -> MSStatement r

emptyStmt :: MSStatement r

multi :: [MSStatement r] -> MSStatement r

class (VariableSym r, StatementSym r) => AssignStatement r where

Methods

(&-=) :: SVariable r -> SValue r -> MSStatement r infixl 1

(&+=) :: SVariable r -> SValue r -> MSStatement r infixl 1

(&++) :: SVariable r -> MSStatement r infixl 8

(&--) :: SVariable r -> MSStatement r infixl 8

assign :: SVariable r -> SValue r -> MSStatement r

(&=) :: AssignStatement r => SVariable r -> SValue r -> MSStatement r infixr 1

assignToListIndex :: (StatementSym r, VariableValue r, List r) => SVariable r -> SValue r -> SValue r -> MSStatement r

class (VariableSym r, StatementSym r, ScopeSym r) => DeclStatement r where

Methods

varDec :: SVariable r -> r (Scope r) -> MSStatement r

varDecDef :: SVariable r -> r (Scope r) -> SValue r -> MSStatement r

listDec :: Integer -> SVariable r -> r (Scope r) -> MSStatement r

listDecDef :: SVariable r -> r (Scope r) -> [SValue r] -> MSStatement r

setDec :: SVariable r -> r (Scope r) -> MSStatement r

setDecDef :: SVariable r -> r (Scope r) -> SValue r -> MSStatement r

arrayDec :: Integer -> SVariable r -> r (Scope r) -> MSStatement r

arrayDecDef :: SVariable r -> r (Scope r) -> [SValue r] -> MSStatement r

constDecDef :: SVariable r -> r (Scope r) -> SValue r -> MSStatement r

funcDecDef :: SVariable r -> r (Scope r) -> [SVariable r] -> MSBody r -> MSStatement r

class (VariableSym r, StatementSym r) => IOStatement r where

Methods

print :: SValue r -> MSStatement r

printLn :: SValue r -> MSStatement r

printStr :: String -> MSStatement r

printStrLn :: String -> MSStatement r

printFile :: SValue r -> SValue r -> MSStatement r

printFileLn :: SValue r -> SValue r -> MSStatement r

printFileStr :: SValue r -> String -> MSStatement r

printFileStrLn :: SValue r -> String -> MSStatement r

getInput :: SVariable r -> MSStatement r

discardInput :: MSStatement r

getFileInput :: SValue r -> SVariable r -> MSStatement r

discardFileInput :: SValue r -> MSStatement r

openFileR :: SVariable r -> SValue r -> MSStatement r

openFileW :: SVariable r -> SValue r -> MSStatement r

openFileA :: SVariable r -> SValue r -> MSStatement r

closeFile :: SValue r -> MSStatement r

getFileInputLine :: SValue r -> SVariable r -> MSStatement r

discardFileLine :: SValue r -> MSStatement r

getFileInputAll :: SValue r -> SVariable r -> MSStatement r

class (VariableSym r, StatementSym r) => StringStatement r where

Methods

stringSplit :: Char -> SVariable r -> SValue r -> MSStatement r

stringListVals :: [SVariable r] -> SValue r -> MSStatement r

stringListLists :: [SVariable r] -> SValue r -> MSStatement r

class ValueSym r => FunctionSym r

Associated Types

type Function r

class (VariableSym r, StatementSym r) => FuncAppStatement r where

Methods

inOutCall :: InOutCall r

extInOutCall :: Library -> InOutCall r

class StatementSym r => CommentStatement r where

Methods

comment :: Comment -> MSStatement r

class (BodySym r, VariableSym r) => ControlStatement r where

Methods

break :: MSStatement r

continue :: MSStatement r

returnStmt :: SValue r -> MSStatement r

throw :: Label -> MSStatement r

ifCond :: [(SValue r, MSBody r)] -> MSBody r -> MSStatement r

String of if-else statements. Arguments: List of predicates and bodies (if this then that), Body for else branch

switch :: SValue r -> [(SValue r, MSBody r)] -> MSBody r -> MSStatement r

ifExists :: SValue r -> MSBody r -> MSBody r -> MSStatement r

for :: MSStatement r -> SValue r -> MSStatement r -> MSBody r -> MSStatement r

forRange :: SVariable r -> SValue r -> SValue r -> SValue r -> MSBody r -> MSStatement r

forEach :: SVariable r -> SValue r -> MSBody r -> MSStatement r

while :: SValue r -> MSBody r -> MSStatement r

tryCatch :: MSBody r -> MSBody r -> MSStatement r

assert :: SValue r -> SValue r -> MSStatement r

ifNoElse :: ControlStatement r => [(SValue r, MSBody r)] -> MSStatement r

switchAsIf :: (ControlStatement r, Comparison r) => SValue r -> [(SValue r, MSBody r)] -> MSBody r -> MSStatement r

class VisibilitySym r where

Associated Types

type Visibility r

Methods

private :: r (Visibility r)

public :: r (Visibility r)

class VariableSym r => ParameterSym r where

Associated Types

type Parameter r

Methods

param :: SVariable r -> MSParameter r

pointerParam :: SVariable r -> MSParameter r

class (BodySym r, ParameterSym r, VisibilitySym r) => MethodSym r where

Associated Types

type Method r

Methods

docMain :: MSBody r -> SMethod r

function :: Label -> r (Visibility r) -> VSType r -> [MSParameter r] -> MSBody r -> SMethod r

mainFunction :: MSBody r -> SMethod r

docFunc :: String -> [String] -> Maybe String -> SMethod r -> SMethod r

inOutFunc :: Label -> r (Visibility r) -> InOutFunc r

docInOutFunc :: Label -> r (Visibility r) -> DocInOutFunc r

convType :: TypeSym r => CodeType -> VSType r