Drasil.Shared.LanguageRenderer.Common

Description

Implementations defined here are valid in some, but not all, language renderers

Documentation

boolRender :: String

bool :: CommonRenderSym r => VSType r

extVar :: CommonRenderSym r => Label -> Label -> VSType r -> SVariable r

funcType :: CommonRenderSym r => [VSType r] -> VSType r -> VSType r

extFuncAppMixedArgs :: CommonRenderSym r => Library -> MixedCall r

listAccessFunc :: CommonRenderSym r => VSType r -> SValue r -> VSFunction r

listSetFunc :: CommonRenderSym r => (Doc -> Doc -> Doc) -> SValue r -> SValue r -> SValue r -> VSFunction r

forEach' :: CommonRenderSym r => (r (Variable r) -> r (Value r) -> r (Body r) -> Doc) -> SVariable r -> SValue r -> MSBody r -> MSStatement r

varDecDef :: CommonRenderSym r => SVariable r -> r (Scope r) -> Maybe (SValue r) -> MSStatement r

listSize :: CommonRenderSym r => SValue r -> SValue r

Call to get the size of a list in a language where this is not a method.