Drasil.Shared.LanguageRenderer.CommonPseudoOO

Description

Implementations defined here are valid in some, but not all, language renderers

Documentation

int :: CommonRenderSym r => VSType r

constructor :: OORenderSym r => Label -> [MSParameter r] -> Initializers r -> MSBody r -> SMethod r

doxFunc :: CommonRenderSym r => String -> [String] -> Maybe String -> SMethod r -> SMethod r

doxClass :: OORenderSym r => String -> SClass r -> SClass r

doxMod :: OORenderSym r => String -> String -> [String] -> String -> SFile r -> SFile r

docMod' :: OORenderSym r => String -> String -> [String] -> String -> SFile r -> SFile r

modDoc' :: ModuleDocRenderer

Generates Markdown/DocC style module doc comment. Useful for Swift, which follows DocC, Julia, which uses Markdown, and any other language that doesn't have Support for a document generator.

functionDoc :: FuncDocRenderer

Generates Markdown/DocC style function doc comment.

extVar :: VariableSym r => Library -> Label -> VSType r -> SVariable r

classVar :: CommonRenderSym r => (Doc -> Doc -> Doc) -> VSType r -> SVariable r -> SVariable r

objVarSelf :: OORenderSym r => SVariable r -> SVariable r

indexOf :: OORenderSym r => Label -> SValue r -> SValue r -> SValue r

contains :: OORenderSym r => Label -> SValue r -> SValue r -> SValue r

containsInt :: OORenderSym r => Label -> Label -> SValue r -> SValue r -> SValue r

listAddFunc :: OORenderSym r => Label -> SValue r -> SValue r -> VSFunction r

discardFileLine :: OORenderSym r => Label -> SValue r -> MSStatement r

intClass :: (OORenderSym r, Monad r) => (Label -> Doc -> Doc -> Doc -> Doc -> Doc) -> Label -> r (Visibility r) -> r ParentSpec -> [CSStateVar r] -> [SMethod r] -> [SMethod r] -> CS (r Doc)

An internal function for creating a class. Parameters: render function, class name, scope, parent, class variables, constructor(s), methods

funcType :: TypeSym r => [VSType r] -> VSType r -> VSType r

buildModule :: OORenderSym r => Label -> FS Doc -> FS Doc -> FS Doc -> [SMethod r] -> [SClass r] -> FSModule r

arrayType :: CommonRenderSym r => VSType r -> VSType r

pi :: CommonRenderSym r => SValue r

printSt :: CommonRenderSym r => SValue r -> SValue r -> MSStatement r

arrayDec :: CommonRenderSym r => SValue r -> SVariable r -> r (Scope r) -> MSStatement r

arrayDecDef :: CommonRenderSym r => SVariable r -> r (Scope r) -> [SValue r] -> MSStatement r

openFileA :: CommonRenderSym r => (SValue r -> VSType r -> SValue r -> SValue r) -> SVariable r -> SValue r -> MSStatement r

forEach :: CommonRenderSym r => Doc -> Doc -> Doc -> Doc -> SVariable r -> SValue r -> MSBody r -> MSStatement r

docMain :: OORenderSym r => MSBody r -> SMethod r

mainFunction :: OORenderSym r => VSType r -> Label -> MSBody r -> SMethod r

buildModule' :: OORenderSym r => Label -> (String -> r (Import r)) -> [Label] -> [SMethod r] -> [SClass r] -> FSModule r

Used by the language renderers to build the module. n is the module name inc is the include is is the import statements ms is the class methods cs is the classes

call' :: CommonRenderSym r => String -> Maybe Library -> Maybe Doc -> MixedCall r

First parameter is language name, rest similar to call from RendererClassesCommon

listSizeFunc :: OORenderSym r => VSFunction r

listAccessFunc' :: OORenderSym r => Label -> VSType r -> SValue r -> VSFunction r

string :: CommonRenderSym r => VSType r

constDecDef :: CommonRenderSym r => SVariable r -> r (Scope r) -> SValue r -> MSStatement r

docInOutFunc :: CommonRenderSym r => ([SVariable r] -> [SVariable r] -> [SVariable r] -> MSBody r -> SMethod r) -> String -> [(String, SVariable r)] -> [(String, SVariable r)] -> [(String, SVariable r)] -> MSBody r -> SMethod r

bindingError :: String -> String

extFuncAppMixedArgs :: ValueExpression r => Library -> MixedCall r

notNull :: CommonRenderSym r => String -> SValue r -> SValue r

listDecDef :: CommonRenderSym r => SVariable r -> r (Scope r) -> [SValue r] -> MSStatement r

destructorError :: String -> String

stateVarDef :: (OORenderSym r, Monad r) => r (Visibility r) -> r (Permanence r) -> SVariable r -> SValue r -> CS (r Doc)

constVar :: (CommonRenderSym r, Monad r) => Doc -> r (Visibility r) -> SVariable r -> SValue r -> CS (r Doc)

litArray :: CommonRenderSym r => (Doc -> Doc) -> VSType r -> [SValue r] -> SValue r

litSet :: CommonRenderSym r => (Doc -> Doc) -> (Doc -> Doc) -> VSType r -> [SValue r] -> SValue r

listSetFunc :: Doc -> Doc -> Doc

litSetFunc :: CommonRenderSym r => String -> VSType r -> [SValue r] -> SValue r

extraClass :: OORenderSym r => Label -> Maybe Label -> [CSStateVar r] -> [SMethod r] -> [SMethod r] -> SClass r

listAccessFunc :: CommonRenderSym r => r (Value r) -> Doc

doubleRender :: String

double :: CommonRenderSym r => VSType r

openFileR :: CommonRenderSym r => (SValue r -> VSType r -> SValue r) -> SVariable r -> SValue r -> MSStatement r

openFileW :: CommonRenderSym r => (SValue r -> VSType r -> SValue r -> SValue r) -> SVariable r -> SValue r -> MSStatement r

stateVar :: (OORenderSym r, Monad r) => r (Visibility r) -> r (Permanence r) -> SVariable r -> CS (r Doc)

self :: OORenderSym r => SVariable r

multiAssign :: CommonRenderSym r => (Doc -> Doc) -> [SVariable r] -> [SValue r] -> MSStatement r

multiReturn :: CommonRenderSym r => (Doc -> Doc) -> [SValue r] -> MSStatement r

listDec :: CommonRenderSym r => SVariable r -> r (Scope r) -> MSStatement r

funcDecDef :: OORenderSym r => SVariable r -> r (Scope r) -> [SVariable r] -> MSBody r -> MSStatement r

inOutCall :: CommonRenderSym r => (Label -> VSType r -> [SValue r] -> SValue r) -> Label -> [SValue r] -> [SVariable r] -> [SVariable r] -> MSStatement r

forLoopError :: String -> String

mainBody :: CommonRenderSym r => MSBody r -> SMethod r

inOutFunc :: CommonRenderSym r => (VSType r -> [MSParameter r] -> MSBody r -> SMethod r) -> [SVariable r] -> [SVariable r] -> [SVariable r] -> MSBody r -> SMethod r

docInOutFunc' :: CommonRenderSym r => FuncDocRenderer -> ([SVariable r] -> [SVariable r] -> [SVariable r] -> MSBody r -> SMethod r) -> String -> [(String, SVariable r)] -> [(String, SVariable r)] -> [(String, SVariable r)] -> MSBody r -> SMethod r

bool :: TypeSym r => VSType r

floatRender :: String

float :: CommonRenderSym r => VSType r

stringRender' :: String

string' :: CommonRenderSym r => VSType r

inherit :: Monad r => Maybe Label -> r ParentSpec

implements :: Monad r => [Label] -> r ParentSpec

listSize :: List r => SValue r -> SValue r

Finds the size of a list. Arguments are: List

setDecDef :: CommonRenderSym r => SVariable r -> r (Scope r) -> [SValue r] -> MSStatement r

setDec :: OORenderSym r => (r (Value r) -> Doc) -> SValue r -> SVariable r -> r (Scope r) -> MSStatement r

listAdd :: CommonRenderSym r => SValue r -> SValue r -> SValue r -> SValue r

Call to insert a value into a list in a language where this is not a method.

listAppend :: CommonRenderSym r => SValue r -> SValue r -> SValue r

Call to append a value to a list in a language where this is not a method.

intToIndex :: SValue r -> SValue r

Convert an integer to an index in a 0-indexed language Since GOOL is 0-indexed, no adjustments need be made

indexToInt :: SValue r -> SValue r

Convert an index to an integer in a 0-indexed language Since GOOL is 0-indexed, no adjustments need be made

intToIndex' :: CommonRenderSym r => SValue r -> SValue r

Convert an integer to an index in a 1-indexed language Since GOOL is 0-indexed, we need to add 1

indexToInt' :: CommonRenderSym r => SValue r -> SValue r

Convert an index to an integer in a 1-indexed language Since GOOL is 0-indexed, we need to subtract 1

varDecDef :: DeclStatement r => SVariable r -> r (Scope r) -> SValue r -> MSStatement r

openFileR' :: CommonRenderSym r => SValue r -> SValue r

openFileW' :: CommonRenderSym r => SValue r -> SValue r

openFileA' :: CommonRenderSym r => SValue r -> SValue r

argExists :: CommonRenderSym r => Integer -> SValue r

global :: Monad r => r ScopeData

setMethodCall :: OORenderSym r => Label -> SValue r -> SValue r -> SValue r