Drasil.Shared.LanguageRenderer.LanguagePolymorphic

Description

Implementations defined here are valid for any language renderer.

Documentation

fileFromData :: OORenderSym r => (FilePath -> r (Module r) -> r (File r)) -> FilePath -> FSModule r -> SFile r

multiBody :: (CommonRenderSym r, Monad r) => [MSBody r] -> MS (r Doc)

block :: (CommonRenderSym r, Monad r) => [MSStatement r] -> MS (r Doc)

multiBlock :: (CommonRenderSym r, Monad r) => [MSBlock r] -> MS (r Doc)

listInnerType :: OORenderSym r => VSType r -> VSType r

obj :: CommonRenderSym r => ClassName -> VSType r

negateOp :: Monad r => VSOp r

csc :: CommonRenderSym r => SValue r -> SValue r

sec :: CommonRenderSym r => SValue r -> SValue r

cot :: CommonRenderSym r => SValue r -> SValue r

equalOp :: Monad r => VSOp r

notEqualOp :: Monad r => VSOp r

greaterOp :: Monad r => VSOp r

greaterEqualOp :: Monad r => VSOp r

lessOp :: Monad r => VSOp r

lessEqualOp :: Monad r => VSOp r

plusOp :: Monad r => VSOp r

minusOp :: Monad r => VSOp r

multOp :: Monad r => VSOp r

divideOp :: Monad r => VSOp r

moduloOp :: Monad r => VSOp r

var :: CommonRenderSym r => Label -> VSType r -> SVariable r

staticVar :: CommonRenderSym r => Label -> VSType r -> SVariable r

objVar :: CommonRenderSym r => SVariable r -> SVariable r -> SVariable r

classVarCheckStatic :: CommonRenderSym r => r (Variable r) -> r (Variable r)

To be used in classVar implementations. Throws an error if the variable is not static since classVar is for accessing static variables from a class

arrayElem :: OORenderSym r => SValue r -> SVariable r -> SVariable r

local :: Monad r => r ScopeData

litChar :: CommonRenderSym r => (Doc -> Doc) -> Char -> SValue r

litDouble :: CommonRenderSym r => Double -> SValue r

litInt :: CommonRenderSym r => Integer -> SValue r

litString :: CommonRenderSym r => String -> SValue r

valueOf :: CommonRenderSym r => SVariable r -> SValue r

arg :: CommonRenderSym r => SValue r -> SValue r -> SValue r

argsList :: CommonRenderSym r => String -> SValue r

call :: CommonRenderSym r => Doc -> Maybe Library -> Maybe Doc -> MixedCall r

First parameter is separator between name and value for named arguments, rest similar to call from RendererClasses

funcAppMixedArgs :: CommonRenderSym r => MixedCall r

selfFuncAppMixedArgs :: CommonRenderSym r => Doc -> SVariable r -> MixedCall r

newObjMixedArgs :: CommonRenderSym r => String -> MixedCtorCall r

lambda :: CommonRenderSym r => ([r (Variable r)] -> r (Value r) -> Doc) -> [SVariable r] -> SValue r -> SValue r

objAccess :: CommonRenderSym r => SValue r -> VSFunction r -> SValue r

objMethodCall :: CommonRenderSym r => Label -> VSType r -> SValue r -> [SValue r] -> NamedArgs r -> SValue r

func :: CommonRenderSym r => Label -> VSType r -> [SValue r] -> VSFunction r

get :: OORenderSym r => SValue r -> SVariable r -> SValue r

set :: OORenderSym r => SValue r -> SVariable r -> SValue r -> SValue r

listAdd :: OORenderSym r => SValue r -> SValue r -> SValue r -> SValue r

listAppend :: OORenderSym r => SValue r -> SValue r -> SValue r

listAccess :: CommonRenderSym r => SValue r -> SValue r -> SValue r

listSet :: CommonRenderSym r => SValue r -> SValue r -> SValue r -> SValue r

getFunc :: OORenderSym r => SVariable r -> VSFunction r

setFunc :: OORenderSym r => VSType r -> SVariable r -> SValue r -> VSFunction r

listAppendFunc :: OORenderSym r => Label -> SValue r -> VSFunction r

stmt :: CommonRenderSym r => MSStatement r -> MSStatement r

loopStmt :: CommonRenderSym r => MSStatement r -> MSStatement r

emptyStmt :: CommonRenderSym r => MSStatement r

assign :: CommonRenderSym r => Terminator -> SVariable r -> SValue r -> MSStatement r

subAssign :: CommonRenderSym r => Terminator -> SVariable r -> SValue r -> MSStatement r

increment :: CommonRenderSym r => SVariable r -> SValue r -> MSStatement r

objDecNew :: OORenderSym r => SVariable r -> r (Scope r) -> [SValue r] -> MSStatement r

print :: CommonRenderSym r => Bool -> Maybe (SValue r) -> SValue r -> SValue r -> MSStatement r

closeFile :: OORenderSym r => Label -> SValue r -> MSStatement r

returnStmt :: CommonRenderSym r => Terminator -> SValue r -> MSStatement r

valStmt :: CommonRenderSym r => Terminator -> SValue r -> MSStatement r

comment :: CommonRenderSym r => Doc -> Label -> MSStatement r

throw :: CommonRenderSym r => (r (Value r) -> Doc) -> Terminator -> Label -> MSStatement r

ifCond :: CommonRenderSym r => (Doc -> Doc) -> Doc -> OptionalSpace -> Doc -> Doc -> Doc -> [(SValue r, MSBody r)] -> MSBody r -> MSStatement r

tryCatch :: CommonRenderSym r => (r (Body r) -> r (Body r) -> Doc) -> MSBody r -> MSBody r -> MSStatement r

construct :: CommonRenderSym r => Label -> MS (r (Type r))

param :: CommonRenderSym r => (r (Variable r) -> Doc) -> SVariable r -> MSParameter r

method :: OORenderSym r => Label -> r (Visibility r) -> r (Permanence r) -> VSType r -> [MSParameter r] -> MSBody r -> SMethod r

getMethod :: OORenderSym r => SVariable r -> SMethod r

setMethod :: OORenderSym r => SVariable r -> SMethod r

initStmts :: OORenderSym r => Initializers r -> MSBody r

function :: OORenderSym r => Label -> r (Visibility r) -> VSType r -> [MSParameter r] -> MSBody r -> SMethod r

docFuncRepr :: CommonRenderSym r => FuncDocRenderer -> String -> [String] -> [String] -> SMethod r -> SMethod r

docFunc :: CommonRenderSym r => FuncDocRenderer -> String -> [String] -> Maybe String -> SMethod r -> SMethod r

buildClass :: OORenderSym r => Maybe Label -> [CSStateVar r] -> [SMethod r] -> [SMethod r] -> SClass r

implementingClass :: OORenderSym r => Label -> [Label] -> [CSStateVar r] -> [SMethod r] -> [SMethod r] -> SClass r

docClass :: OORenderSym r => ClassDocRenderer -> String -> SClass r -> SClass r

commentedClass :: (OORenderSym r, Monad r) => CS (r (BlockComment r)) -> SClass r -> CS (r Doc)

modFromData :: Label -> (Doc -> r (Module r)) -> FS Doc -> FSModule r

fileDoc :: OORenderSym r => String -> (r (Module r) -> r (Block r)) -> r (Block r) -> FSModule r -> SFile r

docMod :: OORenderSym r => ModuleDocRenderer -> String -> String -> [String] -> String -> SFile r -> SFile r

Generates a file for a documented module. mdr is a function that takes description, author, and module name and returns a doc comment e is the file extension d is the description (I think) a is a list of authors dt is the date fl is the file

newtype OptionalSpace

Constructors

OSpace

Fields

oSpace :: Doc

defaultOptSpace :: OptionalSpace

smartAdd :: CommonRenderSym r => SValue r -> SValue r -> SValue r

smartSub :: CommonRenderSym r => SValue r -> SValue r -> SValue r