Drasil.Shared.LanguageRenderer.Macros

Description

Language-polymorphic functions that are defined by GOOL code

Documentation

ifExists :: CommonRenderSym r => SValue r -> MSBody r -> MSBody r -> MSStatement r

decrement1 :: CommonRenderSym r => SVariable r -> MSStatement r

increment :: CommonRenderSym r => SVariable r -> SValue r -> MSStatement r

increment1 :: CommonRenderSym r => SVariable r -> MSStatement r

runStrategy :: (CommonRenderSym r, Monad r) => Label -> [(Label, MSBody r)] -> Maybe (SValue r) -> Maybe (SVariable r) -> MS (r Doc)

listSlice :: CommonRenderSym r => Maybe (SValue r) -> Maybe (SValue r) -> Maybe (SValue r) -> SVariable r -> SValue r -> MSBlock r

makeSetterVal :: CommonRenderSym r => Label -> SValue r -> Maybe Integer -> Maybe (SValue r) -> SValue r -> SValue r -> r (Scope r) -> (MSStatement r, SValue r)

Gets the expression and code for setting bounds in a list slice Input: - String: Variable name for bound (to be created if necessary), - SValue: step value - Maybe Integer: literal value of step, if exists - Maybe SValue: given value of bound - SValue: value of bound if bound not given and step is positive - SValue: value of bound if bound not given and step is negative Output: (MSStatement, SValue): (setter, value) of bound

stringListVals :: CommonRenderSym r => [SVariable r] -> SValue r -> MSStatement r

stringListLists :: CommonRenderSym r => [SVariable r] -> SValue r -> MSStatement r

forRange :: CommonRenderSym r => SVariable r -> SValue r -> SValue r -> SValue r -> MSBody r -> MSStatement r

notifyObservers :: OORenderSym r => VSFunction r -> VSType r -> MSStatement r

notifyObservers' :: OORenderSym r => VSFunction r -> VSType r -> MSStatement r