GOOL.Drasil.LanguageRenderer.Macros

Description

Language-polymorphic functions that are defined by GOOL code

Documentation

ifExists :: RenderSym r => SValue r -> MSBody r -> MSBody r -> MSStatement r

decrement1 :: RenderSym r => SVariable r -> MSStatement r

increment :: RenderSym r => SVariable r -> SValue r -> MSStatement r

increment1 :: RenderSym r => SVariable r -> MSStatement r

runStrategy :: (RenderSym r, Monad r) => Label -> [(Label, MSBody r)] -> Maybe (SValue r) -> Maybe (SVariable r) -> MS (r Doc)

listSlice :: RenderSym r => Maybe (SValue r) -> Maybe (SValue r) -> Maybe (SValue r) -> SVariable r -> SValue r -> MSBlock r

makeSetterVal :: RenderSym r => Label -> SValue r -> Maybe Integer -> Maybe (SValue r) -> SValue r -> SValue r -> (MSStatement r, SValue r)

Gets the expression and code for setting bounds in a list slice Input: - String: Variable name for bound (to be created if necessary), - SValue: step value - Maybe Integer: literal value of step, if exists - Maybe SValue: given value of bound - SValue: value of bound if bound not given and step is positive - SValue: value of bound if bound not given and step is negative Output: (MSStatement, SValue): (setter, value) of bound

stringListVals :: RenderSym r => [SVariable r] -> SValue r -> MSStatement r

stringListLists :: RenderSym r => [SVariable r] -> SValue r -> MSStatement r

forRange :: RenderSym r => SVariable r -> SValue r -> SValue r -> SValue r -> MSBody r -> MSStatement r

notifyObservers :: RenderSym r => VSFunction r -> VSType r -> MSStatement r

notifyObservers' :: RenderSym r => VSFunction r -> VSType r -> MSStatement r