Drasil.GOOL.InterfaceGOOL

Documentation

type GSProgram a = GS (a (Program a))

type SFile a = FS (a (File a))

type FSModule a = FS (a (Module a))

type SClass a = CS (a (Class a))

type CSStateVar a = CS (a (StateVar a))

type Initializers r = [(SVariable r, SValue r)]

class (SharedProg r, ProgramSym r, OOVariableValue r, OODeclStatement r, OOFuncAppStatement r, OOValueExpression r, InternalValueExp r, GetSet r, ObserverPattern r, StrategyPattern r) => OOProg r

class FileSym r => ProgramSym r where

Associated Types

type Program r

Methods

prog :: Label -> Label -> [SFile r] -> GSProgram r

class ModuleSym r => FileSym r where

Associated Types

type File r

Methods

fileDoc :: FSModule r -> SFile r

docMod :: String -> [String] -> String -> SFile r -> SFile r

class ClassSym r => ModuleSym r where

Associated Types

type Module r

Methods

buildModule :: Label -> [Label] -> [SMethod r] -> [SClass r] -> FSModule r

class (OOMethodSym r, StateVarSym r) => ClassSym r where

Associated Types

type Class r

Methods

buildClass :: Maybe Label -> [CSStateVar r] -> [SMethod r] -> [SMethod r] -> SClass r

Main external method for creating a class. Inputs: parent class, variables, constructor(s), methods

extraClass :: Label -> Maybe Label -> [CSStateVar r] -> [SMethod r] -> [SMethod r] -> SClass r

Creates an extra class. Inputs: class name, the rest are the same as buildClass.

implementingClass :: Label -> [Label] -> [CSStateVar r] -> [SMethod r] -> [SMethod r] -> SClass r

Creates a class implementing interfaces. Inputs: class name, interface names, variables, constructor(s), methods

docClass :: String -> SClass r -> SClass r

class TypeSym r => OOTypeSym r where

Methods

obj :: ClassName -> VSType r

class (VariableSym r, OOTypeSym r) => OOVariableSym r where

Methods

staticVar' :: Bool -> Label -> VSType r -> SVariable r

self :: SVariable r

classVar :: VSType r -> SVariable r -> SVariable r

extClassVar :: VSType r -> SVariable r -> SVariable r

objVar :: SVariable r -> SVariable r -> SVariable r

objVarSelf :: SVariable r -> SVariable r

staticVar :: OOVariableSym r => Label -> VSType r -> SVariable r

staticConst :: OOVariableSym r => Label -> VSType r -> SVariable r

($->) :: OOVariableSym r => SVariable r -> SVariable r -> SVariable r infixl 9

class (ValueSym r, OOTypeSym r) => OOValueSym r

class (VariableValue r, OOVariableSym r) => OOVariableValue r

class (ValueExpression r, OOVariableSym r, OOValueSym r) => OOValueExpression r where

Methods

selfFuncAppMixedArgs :: MixedCall r

newObjMixedArgs :: MixedCtorCall r

extNewObjMixedArgs :: Library -> MixedCtorCall r

libNewObjMixedArgs :: Library -> MixedCtorCall r

selfFuncApp :: OOValueExpression r => PosCall r

newObj :: OOValueExpression r => PosCtorCall r

extNewObj :: OOValueExpression r => Library -> PosCtorCall r

libNewObj :: OOValueExpression r => Library -> PosCtorCall r

class (DeclStatement r, OOVariableSym r) => OODeclStatement r where

Methods

objDecDef :: SVariable r -> r (Scope r) -> SValue r -> MSStatement r

objDecNew :: SVariable r -> r (Scope r) -> [SValue r] -> MSStatement r

extObjDecNew :: Library -> SVariable r -> r (Scope r) -> [SValue r] -> MSStatement r

objDecNewNoParams :: OODeclStatement r => SVariable r -> r (Scope r) -> MSStatement r

extObjDecNewNoParams :: OODeclStatement r => Library -> SVariable r -> r (Scope r) -> MSStatement r

class (FuncAppStatement r, OOVariableSym r) => OOFuncAppStatement r where

Methods

selfInOutCall :: InOutCall r

class (ValueSym r, VariableSym r) => GetSet r where

Methods

get :: SValue r -> SVariable r -> SValue r

set :: SValue r -> SVariable r -> SValue r -> SValue r

class ValueSym r => InternalValueExp r where

Methods

objMethodCallMixedArgs' :: Label -> VSType r -> SValue r -> [SValue r] -> NamedArgs r -> SValue r

Generic function for calling a method. Takes the function name, the return type, the object, a list of positional arguments, and a list of named arguments.

objMethodCall :: InternalValueExp r => VSType r -> SValue r -> Label -> [SValue r] -> SValue r

Calling a method. t is the return type of the method, o is the object, f is the method name, and ps is a list of positional arguments.

objMethodCallNamedArgs :: InternalValueExp r => VSType r -> SValue r -> Label -> NamedArgs r -> SValue r

Calling a method with named arguments.

objMethodCallMixedArgs :: InternalValueExp r => VSType r -> SValue r -> Label -> [SValue r] -> NamedArgs r -> SValue r

Calling a method with a mix of positional and named arguments.

objMethodCallNoParams :: InternalValueExp r => VSType r -> SValue r -> Label -> SValue r

Calling a method with no parameters.

class (MethodSym r, PermanenceSym r) => OOMethodSym r where

Methods

method :: Label -> r (Visibility r) -> r (Permanence r) -> VSType r -> [MSParameter r] -> MSBody r -> SMethod r

getMethod :: SVariable r -> SMethod r

setMethod :: SVariable r -> SMethod r

constructor :: [MSParameter r] -> Initializers r -> MSBody r -> SMethod r

inOutMethod :: Label -> r (Visibility r) -> r (Permanence r) -> InOutFunc r

docInOutMethod :: Label -> r (Visibility r) -> r (Permanence r) -> DocInOutFunc r

privMethod :: OOMethodSym r => Label -> VSType r -> [MSParameter r] -> MSBody r -> SMethod r

pubMethod :: OOMethodSym r => Label -> VSType r -> [MSParameter r] -> MSBody r -> SMethod r

initializer :: OOMethodSym r => [MSParameter r] -> Initializers r -> SMethod r

nonInitConstructor :: OOMethodSym r => [MSParameter r] -> MSBody r -> SMethod r

class (VisibilitySym r, PermanenceSym r, VariableSym r) => StateVarSym r where

Associated Types

type StateVar r

Methods

stateVar :: r (Visibility r) -> r (Permanence r) -> SVariable r -> CSStateVar r

stateVarDef :: r (Visibility r) -> r (Permanence r) -> SVariable r -> SValue r -> CSStateVar r

constVar :: r (Visibility r) -> SVariable r -> SValue r -> CSStateVar r

privDVar :: StateVarSym r => SVariable r -> CSStateVar r

pubDVar :: StateVarSym r => SVariable r -> CSStateVar r

pubSVar :: StateVarSym r => SVariable r -> CSStateVar r

class PermanenceSym r where

Associated Types

type Permanence r

Methods

static :: r (Permanence r)

dynamic :: r (Permanence r)

class FunctionSym r => OOFunctionSym r where

Methods

func :: Label -> VSType r -> [SValue r] -> VSFunction r

objAccess :: SValue r -> VSFunction r -> SValue r

($.) :: OOFunctionSym r => SValue r -> VSFunction r -> SValue r infixl 9

selfAccess :: (OOVariableValue r, OOFunctionSym r) => VSFunction r -> SValue r

class (StatementSym r, OOFunctionSym r) => ObserverPattern r where

Methods

notifyObservers :: VSFunction r -> VSType r -> MSStatement r

observerListName :: Label

initObserverList :: DeclStatement r => VSType r -> [SValue r] -> r (Scope r) -> MSStatement r

addObserver :: (StatementSym r, OOVariableValue r, List r) => SValue r -> MSStatement r

class (BodySym r, VariableSym r) => StrategyPattern r where

Methods

runStrategy :: Label -> [(Label, MSBody r)] -> Maybe (SValue r) -> Maybe (SVariable r) -> MSBlock r

convTypeOO :: OOTypeSym r => CodeType -> VSType r