GOOL.Drasil.InterfaceGOOL

Documentation

type VSFunction a = VS (a (Function a))

class (SharedProg r, ProgramSym r, OOVariableValue r, OODeclStatement r, OOFuncAppStatement r, OOValueExpression r, InternalValueExp r, GetSet r, ObserverPattern r, StrategyPattern r) => OOProg r

class TypeSym r => OOTypeSym r where

Methods

obj :: ClassName -> VSType r

class (VariableSym r, OOTypeSym r) => OOVariableSym r where

Methods

staticVar :: Label -> VSType r -> SVariable r

self :: SVariable r

classVar :: VSType r -> SVariable r -> SVariable r

extClassVar :: VSType r -> SVariable r -> SVariable r

objVar :: SVariable r -> SVariable r -> SVariable r

objVarSelf :: SVariable r -> SVariable r

($->) :: OOVariableSym r => SVariable r -> SVariable r -> SVariable r infixl 9

class (ValueSym r, OOTypeSym r) => OOValueSym r

class (VariableValue r, OOVariableSym r) => OOVariableValue r

class (ValueExpression r, OOVariableSym r, OOValueSym r) => OOValueExpression r where

Methods

selfFuncAppMixedArgs :: MixedCall r

newObjMixedArgs :: MixedCtorCall r

extNewObjMixedArgs :: Library -> MixedCtorCall r

libNewObjMixedArgs :: Library -> MixedCtorCall r

selfFuncApp :: OOValueExpression r => PosCall r

newObj :: OOValueExpression r => PosCtorCall r

extNewObj :: OOValueExpression r => Library -> PosCtorCall r

libNewObj :: OOValueExpression r => Library -> PosCtorCall r

class (DeclStatement r, OOVariableSym r) => OODeclStatement r where

Methods

objDecDef :: SVariable r -> SValue r -> MSStatement r

objDecNew :: SVariable r -> [SValue r] -> MSStatement r

extObjDecNew :: Library -> SVariable r -> [SValue r] -> MSStatement r

objDecNewNoParams :: OODeclStatement r => SVariable r -> MSStatement r

extObjDecNewNoParams :: OODeclStatement r => Library -> SVariable r -> MSStatement r

class (FuncAppStatement r, OOVariableSym r) => OOFuncAppStatement r where

Methods

selfInOutCall :: InOutCall r

class (ValueSym r, VariableSym r) => GetSet r where

Methods

get :: SValue r -> SVariable r -> SValue r

set :: SValue r -> SVariable r -> SValue r -> SValue r

class ValueSym r => InternalValueExp r where

Methods

objMethodCallMixedArgs' :: Label -> VSType r -> SValue r -> [SValue r] -> NamedArgs r -> SValue r

Generic function for calling a method. Takes the function name, the return type, the object, a list of positional arguments, and a list of named arguments.

objMethodCall :: InternalValueExp r => VSType r -> SValue r -> Label -> [SValue r] -> SValue r

Calling a method. t is the return type of the method, o is the object, f is the method name, and ps is a list of positional arguments.

objMethodCallNamedArgs :: InternalValueExp r => VSType r -> SValue r -> Label -> NamedArgs r -> SValue r

Calling a method with named arguments.

objMethodCallMixedArgs :: InternalValueExp r => VSType r -> SValue r -> Label -> [SValue r] -> NamedArgs r -> SValue r

Calling a method with a mix of positional and named arguments.

objMethodCallNoParams :: InternalValueExp r => VSType r -> SValue r -> Label -> SValue r

Calling a method with no parameters.

class ValueSym r => FunctionSym r where

Associated Types

type Function r

Methods

func :: Label -> VSType r -> [SValue r] -> VSFunction r

objAccess :: SValue r -> VSFunction r -> SValue r

($.) :: FunctionSym r => SValue r -> VSFunction r -> SValue r infixl 9

selfAccess :: (OOVariableValue r, FunctionSym r) => VSFunction r -> SValue r

class (StatementSym r, FunctionSym r) => ObserverPattern r where

Methods

notifyObservers :: VSFunction r -> VSType r -> MSStatement r

observerListName :: Label

initObserverList :: DeclStatement r => VSType r -> [SValue r] -> MSStatement r

addObserver :: (StatementSym r, OOVariableValue r, List r) => SValue r -> MSStatement r

class (BodySym r, VariableSym r) => StrategyPattern r where

Methods

runStrategy :: Label -> [(Label, MSBody r)] -> Maybe (SValue r) -> Maybe (SVariable r) -> MSBlock r

convTypeOO :: OOTypeSym r => CodeType -> VSType r